[CHANGE] Using `allianceauth-app-utils` to get the absolute URL
This might possibly fix an issue where the Discord bot command is looking for `MUMBLE_URL` to be set.

I can't reproduce the issue at all, so I have to guess here. But I previously used `get_site_url` from `allianceauth-discordbot` for this and that one also has a `mumble_active` check which may or may not trigger this message.